pyproject.toml
[project]
name = "takone"
version = "0.1.0"
description = "Takone — AI-powered video creation agent"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "AGPL-3.0"}
keywords = ["ai", "video", "agent", "generation", "creation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "anthropic>=0.40.0,<1.0.0",
    "openai>=1.50.0,<2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "playwright>=1.40.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
vision = ["opencv-python>=4.8.0"]
all = ["takone[vision]"]

[project.urls]
Homepage = "https://github.com/wormholeportal/takone"
Repository = "https://github.com/wormholeportal/takone"
Issues = "https://github.com/wormholeportal/takone/issues"

[project.scripts]
takone = "agents.director:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["agents*", "core*", "utils*", "skills*"]
